Add Choose service modules

+++ lib/app/data/models/data_service/data_service_model.dart
@@ -20,7 +20,7 @@ class DataService {
   final int idSatuan;
 
   @JsonKey(name: 'harga')
-  final int harga;
+  final double harga;
 
   @JsonKey(name: 'nama_satuan')
   final String namaSatuan;
